Added the canvas to the front-end and linked the buton to request the game session and join to test. There seem to be an issue with csrf token..
