+++ _main_project/a_game/models.py
@@ -9,7 +9,7 @@ class GameSession(models.Model):
 	player2 = models.ForeignKey(Account, on_delete=models.CASCADE, related_name="player2", null=True)
 	player1_score = models.IntegerField(default=0)
 	player2_score = models.IntegerField(default=0)
-	winner = models.ForeignKey(Account, on_delete=models.CASCADE, related_name="winner", null=True)
+	spectators = models.ManyToManyField(Account, related_name="spectators", blank=True)
 	created_at = models.DateTimeField(auto_now_add=True)
 	is_active = models.BooleanField(default=True)
